refactor: split into cargo workspace with core, gates, and language crates
