+++ Cargo.toml
@@ -1,10 +1,3 @@
-[package]
-name = "aspirina"
-version = "0.1.1"
-edition = "2021"
-description = "A modular neural network library with matrix operations and backpropagation"
-license = "MIT"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
+[workspace]
+members = ["core", "gates", "language"]
+resolver = "2"
