+++ core/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod calc;
+pub mod matrix;
