+++ gates/src/computer/assembler.rs
