+++ gates/src/computer/full_adder.rs
