+++ gates/src/computer/half_adder.rs
