+++ gates/src/computer/interpreter.rs
