+++ gates/src/computer/memory.rs
