+++ gates/src/computer/registers.rs
