+++ gates/src/lib.rs
@@ -1,0 +1,14 @@
+pub mod layer;
+pub mod neural_network;
+
+pub mod training {
+    pub mod and_gate;
+    pub mod nand_gate;
+    pub mod nor_gate;
+    pub mod not_gate;
+    pub mod or_gate;
+    pub mod xnor_gate;
+    pub mod xor_gate;
+}
+
+pub mod computer;
