+++ gates/src/training/and_gate.rs
@@ -1,10 +1,10 @@
-use aspirina::layer::Layer;
-use aspirina::matrix::Matrix;
-use aspirina::neural_network::NeuralNetwork;
+use crate::layer::Layer;
+use crate::neural_network::NeuralNetwork;
+use aspirina_core::matrix::Matrix;
 
 pub fn run() {
-    println!("=== OR Gate Training ===");
-    println!("Training neural network to learn OR logic gate...");
+    println!("=== AND Gate Training ===");
+    println!("Training neural network to learn AND logic gate...");
 
     let network = create_network();
 
@@ -15,7 +15,7 @@ pub fn run() {
         vec![1.0, 1.0],
     ]);
 
-    let targets = Matrix::new(vec![vec![0.0, 1.0, 1.0, 1.0]]);
+    let targets = Matrix::new(vec![vec![0.0, 0.0, 0.0, 1.0]]);
 
     // Training loop with progress
     let epochs = 10_000;
@@ -27,20 +27,20 @@ pub fn run() {
     }
 
     println!("\n=== Training Complete ===");
-    println!("Testing OR gate logic:");
+    println!("Testing AND gate logic:");
 
-    // Test all OR combinations
+    // Test all AND combinations
     let test_cases = [
-        (vec![0.0, 0.0], "0 OR 0"),
-        (vec![0.0, 1.0], "0 OR 1"),
-        (vec![1.0, 0.0], "1 OR 0"),
-        (vec![1.0, 1.0], "1 OR 1"),
+        (vec![0.0, 0.0], "0 AND 0"),
+        (vec![0.0, 1.0], "0 AND 1"),
+        (vec![1.0, 0.0], "1 AND 0"),
+        (vec![1.0, 1.0], "1 AND 1"),
     ];
 
     for (input_vals, description) in test_cases {
         let result = network.predict(Matrix::new(vec![input_vals.clone()]));
         let output = result.data[0][0];
-        let expected = if input_vals[0] > 0.5 || input_vals[1] > 0.5 {
+        let expected = if input_vals[0] > 0.5 && input_vals[1] > 0.5 {
             1.0
         } else {
             0.0
@@ -66,14 +66,15 @@ pub fn run() {
 }
 
 fn create_network() -> NeuralNetwork {
-    // OR gate needs at least one input to be high
+    // Better biased network for AND gate
+    // AND gate needs both inputs to be high, so we use positive weights
     let layers = vec![
         Layer::new(Matrix::new(vec![
-            vec![1.0, 1.0],   // Positive weights for disjunction
-            vec![0.7, 0.7],   // Second neuron
-            vec![-0.2, -0.2], // Negative bias neuron
+            vec![0.8, 0.8],   // Both inputs weighted positively
+            vec![0.6, 0.6],   // Second neuron also positive
+            vec![-0.3, -0.3], // Third neuron with negative bias
         ])),
-        Layer::new(Matrix::new(vec![vec![1.5, 1.0, -0.3]])),
+        Layer::new(Matrix::new(vec![vec![1.2, 0.8, -0.5]])), // Output layer biased towards conjunction
     ];
 
     NeuralNetwork::new(layers)
