+++ gates/src/training/or_gate.rs
@@ -1,10 +1,10 @@
-use aspirina::layer::Layer;
-use aspirina::matrix::Matrix;
-use aspirina::neural_network::NeuralNetwork;
+use crate::layer::Layer;
+use crate::neural_network::NeuralNetwork;
+use aspirina_core::matrix::Matrix;
 
 pub fn run() {
-    println!("=== NAND Gate Training ===");
-    println!("Training neural network to learn NAND logic gate...");
+    println!("=== OR Gate Training ===");
+    println!("Training neural network to learn OR logic gate...");
 
     let network = create_network();
 
@@ -15,7 +15,7 @@ pub fn run() {
         vec![1.0, 1.0],
     ]);
 
-    let targets = Matrix::new(vec![vec![1.0, 1.0, 1.0, 0.0]]);
+    let targets = Matrix::new(vec![vec![0.0, 1.0, 1.0, 1.0]]);
 
     // Training loop with progress
     let epochs = 10_000;
@@ -27,23 +27,23 @@ pub fn run() {
     }
 
     println!("\n=== Training Complete ===");
-    println!("Testing NAND gate logic:");
+    println!("Testing OR gate logic:");
 
-    // Test all NAND combinations
+    // Test all OR combinations
     let test_cases = [
-        (vec![0.0, 0.0], "0 NAND 0"),
-        (vec![0.0, 1.0], "0 NAND 1"),
-        (vec![1.0, 0.0], "1 NAND 0"),
-        (vec![1.0, 1.0], "1 NAND 1"),
+        (vec![0.0, 0.0], "0 OR 0"),
+        (vec![0.0, 1.0], "0 OR 1"),
+        (vec![1.0, 0.0], "1 OR 0"),
+        (vec![1.0, 1.0], "1 OR 1"),
     ];
 
     for (input_vals, description) in test_cases {
         let result = network.predict(Matrix::new(vec![input_vals.clone()]));
         let output = result.data[0][0];
-        let expected = if input_vals[0] > 0.5 && input_vals[1] > 0.5 {
-            0.0 // NAND is NOT AND
-        } else {
+        let expected = if input_vals[0] > 0.5 || input_vals[1] > 0.5 {
             1.0
+        } else {
+            0.0
         };
 
         println!(
@@ -66,14 +66,14 @@ pub fn run() {
 }
 
 fn create_network() -> NeuralNetwork {
-    // NAND gate is inverted AND - high unless both inputs are high
+    // OR gate needs at least one input to be high
     let layers = vec![
         Layer::new(Matrix::new(vec![
-            vec![-0.8, -0.8], // Negative weights for NAND logic
-            vec![0.5, 0.5],   // Positive bias neuron
-            vec![-0.3, 0.3],  // Mixed weights for complexity
+            vec![1.0, 1.0],   // Positive weights for disjunction
+            vec![0.7, 0.7],   // Second neuron
+            vec![-0.2, -0.2], // Negative bias neuron
         ])),
-        Layer::new(Matrix::new(vec![vec![-1.2, 1.5, 0.8]])),
+        Layer::new(Matrix::new(vec![vec![1.5, 1.0, -0.3]])),
     ];
 
     NeuralNetwork::new(layers)
