+++ language/src/lib.rs
@@ -1,0 +1,1 @@
+pub use aspirina_core;
